+++ www/src/render.ts
@@ -19,7 +19,7 @@ import { Camera } from "./camera.js";
 import { InHandOverlay } from "./ui/inhand_overlay.js";
 import { Environment, PRESET_NAMES } from "./environment.js";
 import GeometryTerrain from "./geometry_terrain.js";
-import {BLEND_MODES, GlobalUniformGroup, LightUniformGroup} from "./renders/BaseRenderer.js";
+import {BLEND_MODES} from "./renders/BaseRenderer.js";
 import { CubeSym } from "./core/CubeSym.js";
 import {
     DEFAULT_CLOUD_HEIGHT,
@@ -95,8 +95,6 @@ export class Renderer {
     settings:               any
     videoCardInfoCache:     any
     options:                any
-    globalUniforms:         GlobalUniformGroup;
-    lightUniforms:          LightUniformGroup;
     defaultShader:          any
     defaultFluidShader:     any
     viewportWidth:          any
@@ -147,6 +145,14 @@ export class Renderer {
 
         //
         this.drop_item_meshes = Array(4096); // new Map();
+    }
+
+    get globalUniforms() {
+        return this.renderBackend.globalUniforms;
+    }
+
+    get lightUniforms() {
+        return this.renderBackend.lightUniforms;
     }
 
     //
@@ -218,9 +224,6 @@ export class Renderer {
         // Init shaders for all resource packs
         await BLOCK.resource_pack_manager.initShaders(renderBackend);
         await BLOCK.resource_pack_manager.initTextures(renderBackend, settings);
-
-        this.globalUniforms = renderBackend.globalUniforms;
-        this.lightUniforms = renderBackend.lightUniforms;
 
         // Make materials for all shaders
         for(let rp of BLOCK.resource_pack_manager.list.values()) {
